Migrate the package building directory out from in port-build to the build directory (${CUR_DIR}../pkgbuild/).
Make X_BUILD_BASE_CFG a requirement now as it's actually a useful
profile style name for the target profile group we're building
code for.  It's used in various scripts anyway, so we should make
it a top-level configuration variable (and likely rename it later)
and then migrate those paths into cfg.sh.